+++ src/main/java/omegadrive/bus/z80/Sg1000Bus.java
@@ -43,8 +43,8 @@ public class Sg1000Bus extends DeviceAwareBus<Tms9918aVdp, TwoButtonsJoypad> imp
     private static final int ROM_SIZE = ROM_END + 1; //48kb
 
     @Override
-    public long read(long addressL, Size size) {
-        int address = (int) addressL;
+    public int read(int addressL, Size size) {
+        int address = addressL;
         if (size != Size.BYTE) {
             LOG.error("Unexpected read, addr : {} , size: {}", address, size);
             return 0xFF;
@@ -60,9 +60,9 @@ public class Sg1000Bus extends DeviceAwareBus<Tms9918aVdp, TwoButtonsJoypad> imp
     }
 
     @Override
-    public void write(long address, long data, Size size) {
+    public void write(int address, int data, Size size) {
         address &= RAM_SIZE - 1;
-        memoryProvider.writeRamByte((int) address, (int) (data & 0xFF));
+        memoryProvider.writeRamByte(address, (data & 0xFF));
     }
 
     /**
